+++ decidim-module-elections_census/app/controllers/decidim/elections_census/about_controller.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ElectionsCensus
+    class AboutController < ApplicationController
+      layout "decidim/elections_voting"
+
+      def show
+      end
+    end
+  end
+end
