+++ decidim-module-elections_census/app/views/decidim/elections_census/booth/authentication/new.html.erb
@@ -1,0 +1,52 @@
+  <div class="row collapse">
+    <div class="row collapse">
+      <div class="columns large-8 large-centered text-center page-title">
+        <h1><%= t(".title") %></h1>
+        <p>
+        <%= t(".subtitle") %>
+        </p>
+      </div>
+    </div>
+    <div class="row">
+      <div class="columns large-6 medium-10 medium-centered">
+        <p><%= t(".form_instructions_1") %></p>
+        <p><%= t(".form_instructions_2") %></p>
+        <p><%= t(".form_instructions_3_html") %></p>
+        <%= decidim_form_for(@form, url: booth_auth_path, method: :post) do |f| %>
+          <%= invisible_captcha %>
+          <div class="card">
+            <div class="card__content">
+              <%= form_required_explanation %>
+
+              <div class="field">
+                <%= f.select :document_type, Decidim::ElectionsCensus::VoterForm::DOCUMENT_TYPES.map {|option| [t(option, scope: "decidim.elections_census.document_types"), option]}, prompt: t("decidim.elections_census.voters.new.select_document_type") %>
+              </div>
+
+              <div class="field">
+                <%= f.text_field :document_number %>
+              </div>
+
+              <div class="field voter-birthday">
+                <%= f.date_select :birthday, start_year: Time.current.year, end_year: 110.years.ago.year, default: 35.years.ago, prompt: { day: t("decidim.elections_census.voters.new.date_select.day"), month: t("decidim.elections_census.voters.new.date_select.month"), year: t("decidim.elections_census.voters.new.date_select.year") } %>
+              </div>
+
+              <div class="field">
+                <%= f.text_field :address_postal_code %>
+              </div>
+
+              <div class="field">
+                <%= f.text_field :mobile_phone_number %>
+              </div>
+            </div>
+          </div>
+          <div class="card">
+            <div class="card__content">
+              <div class="actions">
+                <%= f.submit t("decidim.elections_census.voters.new.send"), class: "button expanded" %>
+              </div>
+            </div>
+          </div>
+        <% end %>
+      </div>
+    </div>
+  </div>
