+++ decidim-module-elections_census/app/views/decidim/elections_census/booth/vote/_vote_options.html.erb
