+++ decidim-module-elections_census/app/views/decidim/elections_census/booth/vote/create.html.erb
@@ -10,7 +10,7 @@
       </p>
       <p>
       <%= t(".verify_vote") %>
-      <%= link_to t(".verify_vote_link"), check_voting_path(receipt: @vote.receipt) %>.
+      <%= link_to t(".verify_vote_link"), verify_path %>.
       </p>
     </div>
