+++ decidim-module-elections_census/app/views/decidim/elections_census/booth/vote/preview.html.erb
@@ -7,7 +7,7 @@
       <% @form.votes.each do |disability, votes| %>
         <li class="accordion-item evote__preview-question is-active" data-accordion-item>
           <a href="#" class="accordion-title flex--sbc">
-            <%= t("decidim.elections_census.voting.vote_options.voting_title", disability_title: t("decidim.elections_census.voting.vote_options.voting_disability_title.#{disability}")) %>
+            <%= t("decidim.elections_census.booth.vote.vote_options.voting_title", disability_title: t("decidim.elections_census.booth.vote.vote_options.voting_disability_title.#{disability}")) %>
           </a>
 
           <div class="accordion-content" data-tab-content>
@@ -27,11 +27,10 @@
       <% end %>
     </ul>
 
-
     <div class="card">
       <div class="card__content">
         <div class="actions">
-          <%= decidim_form_for(@form, url: vote_voting_path(edit_vote: true)) do |f| %>
+          <%= decidim_form_for(@form, url: booth_vote_path) do |f| %>
             <%= f.hidden_field :preview, value: false %>
             <%= f.hidden_field :voting_digest %>
             <%= f.hidden_field :voter_id %>
@@ -44,7 +43,7 @@
             <% end %>
             <%= f.submit t(".edit"), class: "button expanded secondary", name: "edit_vote" %>
           <% end %>
-          <%= decidim_form_for(@form, url: vote_voting_path) do |f| %>
+          <%= decidim_form_for(@form, url: booth_cast_path) do |f| %>
             <%= f.hidden_field :preview, value: false %>
             <%= f.hidden_field :voting_digest %>
             <%= f.hidden_field :voter_id %>
