+++ decidim-module-elections_census/app/views/decidim/elections_census/verify/new.html.erb
@@ -10,7 +10,7 @@
     </div>
     <div class="row">
       <div class="columns large-6 medium-10 medium-centered">
-        <%= decidim_form_for(@form, url: check_voting_path) do |f| %>
+        <%= decidim_form_for(@form, url: verify_path) do |f| %>
           <div class="card">
             <div class="card__content">
               <div class="field">
