+++ decidim-module-elections_census/lib/decidim/elections_census/engine.rb
@@ -14,11 +14,18 @@ module Decidim
           match "search" => "voters#search", as: :search, via: [:get, :post], on: :collection
         end
 
-        resource :voting, only: [:index], as: :voting, controller: :voting do
-          match "/" => "voting#index", as: :elections, via: [:get], on: :collection
-          match "vote" => "voting#vote", as: :vote, via: [:get, :post], on: :collection
-          match "check" => "voting#check", as: :check, via: [:get, :post], on: :collection
+        namespace :booth do
+          match "auth" => "authentication#new", via: [:get]
+          match "auth" => "authentication#create", via: [:post]
+
+          match "vote" => "vote#new", via: [:post]
+          match "preview" => "vote#preview", via: [:post]
+          match "cast" => "vote#create", via: [:post]
         end
+
+        match "about" => "about#show", as: :elections, via: [:get]
+        match "verify" => "verify#new", via: [:get]
+        match "verify" => "verify#create", via: [:post]
 
         root to: "voters#index"
       end
@@ -31,9 +38,9 @@ module Decidim
         if app.secrets.voting_active
           Decidim.menu :menu do |menu|
             menu.item I18n.t("menu.voting", scope: "decidim"),
-              decidim_elections_census.elections_voting_path,
+              decidim_elections_census.elections_path,
               position: 3,
-              active: %r{^/elections\/voting/}
+              active: %r{^/elections/}
           end
         end
         if app.secrets.census_active
